@@ -4,10 +4,12 @@ namespace App\Http\Helpers;
 
 use App\Constants\GlobalConst;
 use App\Models\Admin\ReloadlyApi;
+use App\Services\Monitoring\DomainInstrumentation;
 use Exception;
 use Illuminate\Http\Client\RequestException;
 use Illuminate\Http\Client\Response;
 use Illuminate\Support\Facades\Http;
+use Throwable;
 
 class AirtimeHelper{
 
@@ -48,8 +50,13 @@ class AirtimeHelper{
         "RANGE" => "RANGE",
     ];
 
-    public function __construct()
+    protected DomainInstrumentation $instrumentation;
+
+    protected string $domain = 'topup';
+
+    public function __construct(?DomainInstrumentation $instrumentation = null)
     {
+        $this->instrumentation = $instrumentation ?: app(DomainInstrumentation::class);
         $this->api = ReloadlyApi::reloadly()->mobileTopUp()->first();
         $this->setConfig();
         $this->accessToken();
@@ -145,6 +152,12 @@ class AirtimeHelper{
         $base_url = $this->config['request_url'];
 
         $request_endpoint = $base_url .'/countries'.($iso? '/'.$iso:'');
+        $config = config($this->domain, []);
+        $context = $this->instrumentation->startOperation($this->domain, 'fetch_countries', $config, [
+            'provider' => $this->api->provider ?? data_get($config, 'credentials.provider', 'unknown'),
+            'iso' => $iso,
+        ]);
+
         try{
             $response = Http::withHeaders([
                 'Authorization' => "Bearer " . $access_token,
@@ -159,9 +172,18 @@ class AirtimeHelper{
                 'message' => $error_response['message']??'',
                 'errorCode' => $error_response['errorCode']??'',
             ];
+            $this->instrumentation->recordFailure($context, $e, $data);
             return $data;
+        } catch (Throwable $e) {
+            $this->instrumentation->recordFailure($context, $e);
+            throw $e;
         }
-        if(!is_array($response)) throw new Exception(__("Something went wrong! Please try again."));
+        if(!is_array($response)) {
+            $exception = new Exception(__("Something went wrong! Please try again."));
+            $this->instrumentation->recordFailure($context, $exception);
+            throw $exception;
+        }
+        $this->instrumentation->recordSuccess($context, ['countries' => is_countable($response) ? count($response) : null]);
         return $response;
     }
     /**
@@ -173,6 +195,12 @@ class AirtimeHelper{
         $access_token = $this->access_token;
         $base_url = $this->config['request_url'];
         $request_endpoint = $base_url . "/operators/auto-detect/phone/$phone/country-code/".$iso."?&suggestedAmountsMap=true";
+        $config = config($this->domain, []);
+        $context = $this->instrumentation->startOperation($this->domain, 'auto_detect_operator', $config, [
+            'provider' => $this->api->provider ?? data_get($config, 'credentials.provider', 'unknown'),
+            'phone' => $phone,
+            'iso' => $iso,
+        ]);
         try{
             $response = Http::withHeaders([
                 'Authorization' => "Bearer " . $access_token,
@@ -187,9 +215,18 @@ class AirtimeHelper{
                 'message' => $error_response['message']??'',
                 'errorCode' => $error_response['errorCode']??'',
             ];
+            $this->instrumentation->recordFailure($context, $e, $data);
             return $data;
+        } catch (Throwable $e) {
+            $this->instrumentation->recordFailure($context, $e);
+            throw $e;
         }
-        if(!is_array($response)) throw new Exception(__("Something went wrong! Please try again."));
+        if(!is_array($response)) {
+            $exception = new Exception(__("Something went wrong! Please try again."));
+            $this->instrumentation->recordFailure($context, $exception);
+            throw $exception;
+        }
+        $this->instrumentation->recordSuccess($context, ['operator_id' => $response['operatorId'] ?? null]);
         return $response;
     }
     /**
@@ -201,8 +238,15 @@ class AirtimeHelper{
 
         $base_url = $this->config['request_url'];
         $endpoint = $base_url . "/topups";
+        $config = config($this->domain, []);
 
-        try{
+        $context = $this->instrumentation->startOperation($this->domain, 'make_topup', $config, [
+            'provider' => $this->api->provider ?? data_get($config, 'credentials.provider', 'unknown'),
+            'amount'   => data_get($data, 'amount'),
+            'recipient' => data_get($data, 'recipientPhone') ?? data_get($data, 'recipientEmail'),
+        ]);
+
+        try {
             $response = Http::withHeaders([
                 "Accept: application/com.reloadly.topups-v1+json",
                 "Authorization" => "Bearer " . $this->access_token,
@@ -210,17 +254,30 @@ class AirtimeHelper{
             ])->post($endpoint, $data)->throw(function(Response $response, RequestException $exception) {
                 // throw new Exception($exception->getMessage());
             })->json();
-        }catch(RequestException $e){
+        } catch (RequestException $e) {
             $error_response = json_decode($e->response->body(), true);
-            $data = [
+            $payload = [
                 'status' => false,
                 'message' => $error_response['message']??'',
                 'errorCode' => $error_response['errorCode']??'',
             ];
-            return $data;
+            $this->instrumentation->recordFailure($context, $e, $payload);
+            return $payload;
+        } catch (Throwable $e) {
+            $this->instrumentation->recordFailure($context, $e);
+            throw $e;
         }
 
-        if(!is_array($response)) throw new Exception(__("Something went wrong! Please try again."));
+        if(!is_array($response)) {
+            $exception = new Exception(__("Something went wrong! Please try again."));
+            $this->instrumentation->recordFailure($context, $exception);
+            throw $exception;
+        }
+
+        $this->instrumentation->recordSuccess($context, [
+            'status' => $response['status'] ?? null,
+            'transaction_id' => $response['transactionId'] ?? null,
+        ]);
 
         return $response;
     }
