@@ -2,11 +2,13 @@
 namespace App\Http\Helpers;
 
 use App\Constants\PaymentGatewayConst;
+use App\Contracts\RegionalPaymentProviderInterface;
 use App\Models\Admin\Currency;
 use App\Models\Admin\PaymentGateway as PaymentGatewayModel;
 use App\Models\Admin\PaymentGatewayCurrency;
 use App\Models\TemporaryData;
 use App\Models\Transaction;
+use App\Services\Payments\Regional\RegionalPaymentManager;
 use App\Traits\PaymentGateway\Paypal;
 use App\Traits\PaymentGateway\Stripe;
 use App\Traits\PaymentGateway\Manual;
@@ -25,11 +27,14 @@ use Illuminate\Support\Facades\Route;
 use App\Traits\PaymentGateway\Tatum;
 use App\Traits\PaymentGateway\PerfectMoney;
 use App\Traits\PaymentGateway\PaystackTrait;
+use App\Services\Payments\PaymentProviderInterface;
+use App\Services\Payments\PaymentProviderResolver;
 
 class PaymentGateway {
 
     use Paypal,Stripe,Manual,FlutterwaveTrait,RazorTrait,PagaditoTrait,SslcommerzTrait,CoinGate,Tatum,PerfectMoney,PaystackTrait;
 
+    protected PaymentProviderResolver $providerResolver;
     protected $request_data;
     protected $output;
     protected $currency_input_name = "currency";
@@ -38,14 +43,73 @@ class PaymentGateway {
     protected $predefined_guard;
     protected $predefined_user;
 
+    protected ?RegionalPaymentManager $regionalManager = null;
 
-    public function __construct(array $request_data)
+
+    public function __construct(array $request_data, ?RegionalPaymentManager $regionalManager = null)
     {
         $this->request_data = $request_data;
+        $this->regionalManager = $regionalManager;
     }
 
     public static function init(array $data) {
-        return new PaymentGateway($data);
+        /** @var self $instance */
+        $instance = app(self::class);
+        return $instance->setRequestData($data);
+    }
+
+    public function setRequestData(array $data): self
+    {
+        $this->request_data = $data;
+        return $this;
+    }
+
+    protected function getRegionalManager(): ?RegionalPaymentManager
+    {
+        if ($this->regionalManager) {
+            return $this->regionalManager;
+        }
+
+        if (function_exists('app') && app()->bound(RegionalPaymentManager::class)) {
+            $this->regionalManager = app(RegionalPaymentManager::class);
+        }
+
+        return $this->regionalManager;
+    }
+
+    protected function resolveRegionalProvider(?string $currencyCode): ?RegionalPaymentProviderInterface
+    {
+        $manager = $this->getRegionalManager();
+
+        if (!$currencyCode || !$manager) {
+            return null;
+        }
+
+        return $manager->resolveByCurrency($currencyCode);
+    }
+
+    protected function getRegionalManager(): ?RegionalPaymentManager
+    {
+        if ($this->regionalManager) {
+            return $this->regionalManager;
+        }
+
+        if (function_exists('app') && app()->bound(RegionalPaymentManager::class)) {
+            $this->regionalManager = app(RegionalPaymentManager::class);
+        }
+
+        return $this->regionalManager;
+    }
+
+    protected function resolveRegionalProvider(?string $currencyCode): ?RegionalPaymentProviderInterface
+    {
+        $manager = $this->getRegionalManager();
+
+        if (!$currencyCode || !$manager) {
+            return null;
+        }
+
+        return $manager->resolveByCurrency($currencyCode);
     }
 
     public function gateway() {
@@ -64,6 +128,20 @@ class PaymentGateway {
         if(!$user_wallet) {
             throw ValidationException::withMessages([
                 $this->currency_input_name = __("User wallet not found!"),
+            ]);
+        }
+
+        $regionalProvider = $this->resolveRegionalProvider($gateway_currency->currency_code ?? null);
+        if ($regionalProvider) {
+            $this->output['regional_provider'] = get_class($regionalProvider);
+            $this->output['regional_checkout'] = $regionalProvider->prepareCheckout([
+                'wallet' => $user_wallet,
+                'amount' => (float) ($request_data[$this->amount_input] ?? 0),
+                'currency' => $gateway_currency->currency_code ?? null,
+                'meta' => [
+                    'gateway_alias' => $gateway_currency->gateway->alias ?? null,
+                    'context' => request()->expectsJson() ? 'api' : 'panel',
+                ],
             ]);
         }
 
@@ -86,6 +164,44 @@ class PaymentGateway {
         // limit validation
         $this->limitValidation($this->output);
         return $this;
+    }
+
+    public function executeRegionalPayment(string $currencyCode, array $payload = []): array
+    {
+        $provider = $this->resolveRegionalProvider($currencyCode);
+
+        if (!$provider) {
+            throw new Exception(__('Regional payment provider not available for the selected currency.'));
+        }
+
+        if (!isset($payload['wallet']) && isset($this->output['wallet'])) {
+            $payload['wallet'] = $this->output['wallet'];
+        }
+
+        if (!isset($payload['amount']) && isset($this->output['amount']) && is_object($this->output['amount']) && isset($this->output['amount']->requested_amount)) {
+            $payload['amount'] = (float) $this->output['amount']->requested_amount;
+        }
+
+        return $provider->executePayment($payload);
+    }
+
+    public function refundRegionalPayment(string $currencyCode, array $payload = []): array
+    {
+        $provider = $this->resolveRegionalProvider($currencyCode);
+
+        if (!$provider) {
+            throw new Exception(__('Regional payment provider not available for the selected currency.'));
+        }
+
+        if (!isset($payload['wallet']) && isset($this->output['wallet'])) {
+            $payload['wallet'] = $this->output['wallet'];
+        }
+
+        if (!isset($payload['amount']) && isset($this->output['amount']) && is_object($this->output['amount']) && isset($this->output['amount']->requested_amount)) {
+            $payload['amount'] = (float) $this->output['amount']->requested_amount;
+        }
+
+        return $provider->refundPayment($payload);
     }
 
     public function validator($data) {
@@ -152,15 +268,20 @@ class PaymentGateway {
         if(!$gateway) $gateway = $this->output['gateway'];
         $alias = Str::lower($gateway->alias);
         if($gateway->type == PaymentGatewayConst::AUTOMATIC){
-            $method = PaymentGatewayConst::register($alias);
+            $serviceId = PaymentGatewayConst::register($alias);
         }elseif($gateway->type == PaymentGatewayConst::MANUAL){
-            $method = PaymentGatewayConst::register(strtolower($gateway->type));
+            $serviceId = PaymentGatewayConst::register(strtolower($gateway->type));
         }
 
-        if(method_exists($this,$method)) {
-            return $method;
+        if(!$serviceId) {
+            throw new Exception("Gateway(".$gateway->name.") provider is not registered");
         }
-        return throw new Exception("Gateway(".$gateway->name.") Trait or Method (".$method."()) does not exists");
+
+        $provider = $this->providerResolver->resolve($serviceId);
+        $this->output['provider'] = $provider;
+        $this->output['provider_method'] = $provider->defaultInitializeMethod();
+
+        return $provider;
     }
     public function amount() {
         $currency = $this->output['currency'] ?? null;
@@ -246,8 +367,15 @@ class PaymentGateway {
             }
         }
 
-        $distributeMethod = $this->output['distribute'];
-        return $this->$distributeMethod($output) ?? throw new Exception(__("Something went wrong! Please try again."));
+        $provider = $this->output['provider'] ?? null;
+        if(!$provider instanceof PaymentProviderInterface) {
+            throw new Exception(__('Payment provider not available.'));
+        }
+
+        return $provider->initialize($this, [
+            'output' => $output,
+            'method' => $this->output['provider_method'] ?? null,
+        ]);
     }
     public function authenticateTempData()
     {
@@ -316,47 +444,14 @@ class PaymentGateway {
         $this->gateway();
         $this->output['tempData'] = $tempData;
 
-        if($type == 'flutterWave'){
-            if(method_exists(FlutterwaveTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'razorpay'){
-            if(method_exists(RazorTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'pagadito'){
-            if(method_exists(PagaditoTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'stripe'){
-            if(method_exists(Stripe::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'sslcommerz'){
-            if(method_exists(SslcommerzTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'coingate'){
-            if(method_exists(CoinGate::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'tatum'){
-            if(method_exists(TATUM::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'perfect-money'){
-            if(method_exists(PerfectMoney::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'paystack'){
-            if(method_exists(PaystackTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }else{
-            if(method_exists(Paypal::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
+        $provider = $this->output['provider'] ?? null;
+        if($provider instanceof PaymentProviderInterface) {
+            return $provider->capture($this, [
+                'method' => $method_name,
+                'payload' => $this->output,
+            ]);
         }
+
         throw new Exception("Response method ".$method_name."() does not exists.");
     }
 
