@@ -25,11 +25,14 @@ use App\Traits\PaymentGateway\Tatum;
 use App\Models\Admin\PaymentGateway as PaymentGatewayModel;
 use App\Traits\PaymentGateway\PerfectMoney;
 use App\Traits\PaymentGateway\PaystackTrait;
+use App\Services\Payments\PaymentProviderInterface;
+use App\Services\Payments\PaymentProviderResolver;
 
 class PaymentGatewayApi {
 
     use Paypal,Stripe,Manual,FlutterwaveTrait,RazorTrait,PagaditoTrait,SslcommerzTrait,CoinGate,Tatum,PerfectMoney,PaystackTrait;
 
+    protected PaymentProviderResolver $providerResolver;
     protected $request_data;
     protected $output;
     protected $currency_input_name = "currency";
@@ -39,13 +42,22 @@ class PaymentGatewayApi {
     protected $predefined_user;
 
 
-    public function __construct(array $request_data)
+    public function __construct(PaymentProviderResolver $providerResolver)
     {
-        $this->request_data = $request_data;
+        $this->providerResolver = $providerResolver;
+        $this->request_data = [];
     }
 
     public static function init(array $data) {
-        return new PaymentGatewayApi($data);
+        /** @var self $instance */
+        $instance = app(self::class);
+        return $instance->setRequestData($data);
+    }
+
+    public function setRequestData(array $data): self
+    {
+        $this->request_data = $data;
+        return $this;
     }
 
     public function gateway() {
@@ -157,17 +169,21 @@ class PaymentGatewayApi {
         if(!$gateway) $gateway = $this->output['gateway'];
         $alias = Str::lower($gateway->alias);
         if($gateway->type == PaymentGatewayConst::AUTOMATIC){
-            $method = PaymentGatewayConst::register($alias);
+            $serviceId = PaymentGatewayConst::register($alias);
         }elseif($gateway->type == PaymentGatewayConst::MANUAL){
-            $method = PaymentGatewayConst::register(strtolower($gateway->type));
+            $serviceId = PaymentGatewayConst::register(strtolower($gateway->type));
         }
 
-        if(method_exists($this,$method)) {
-            return $method;
+        if(!$serviceId) {
+            $error = ['error'=>["Gateway(".$gateway->name.") provider is not registered"]];
+            return Helpers::error($error);
         }
 
-        $error = ['error'=>["Gateway(".$gateway->name.") Trait or Method (".$method."()) does not exists"]];
-        return Helpers::error($error);
+        $provider = $this->providerResolver->resolve($serviceId);
+        $this->output['provider'] = $provider;
+        $this->output['provider_method'] = $provider->defaultInitializeMethod();
+
+        return $provider;
     }
     public function amount() {
         $currency = $this->output['currency'] ?? null;
@@ -261,8 +277,16 @@ class PaymentGatewayApi {
             }
         }
 
-        $distributeMethod = $this->output['distribute'];
-        return $this->$distributeMethod($output);
+        $provider = $this->output['provider'] ?? null;
+        if(!$provider instanceof PaymentProviderInterface) {
+            $error = ['error'=>[__('Payment provider not available.')]];
+            return Helpers::error($error);
+        }
+
+        return $provider->initialize($this, [
+            'output' => $output,
+            'method' => $this->output['provider_method'] ?? null,
+        ]);
     }
     public function authenticateTempData()
     {
@@ -338,46 +362,12 @@ class PaymentGatewayApi {
         $this->request_data = $validator_data;
         $this->gateway();
         $this->output['tempData'] = $tempData;
-        if($type == 'flutterWave'){
-            if(method_exists(FlutterwaveTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'razorpay'){
-            if(method_exists(RazorTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'pagadito'){
-            if(method_exists(PagaditoTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'stripe'){
-            if(method_exists(Stripe::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'sslcommerz'){
-            if(method_exists(SslcommerzTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'coingate'){
-            if(method_exists(CoinGate::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'tatum'){
-            if(method_exists(TATUM::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'perfect-money'){
-            if(method_exists(PerfectMoney::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }elseif($type == 'paystack'){
-            if(method_exists(PaystackTrait::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
-        }else{
-            if(method_exists(Paypal::class,$method_name)) {
-                return $this->$method_name($this->output);
-            }
+        $provider = $this->output['provider'] ?? null;
+        if($provider instanceof PaymentProviderInterface) {
+            return $provider->capture($this, [
+                'method' => $method_name,
+                'payload' => $this->output,
+            ]);
         }
 
         $error = ['error'=>["Response method ".$method_name."() does not exists."]];
@@ -391,11 +381,22 @@ class PaymentGatewayApi {
     }
     public function api() {
         $output = $this->output;
-        $output['distribute']   = $this->gatewayDistribute() . "Api";
-        $method = $output['distribute'];
-        $response = $this->$method($output);
+        $provider = $this->output['provider'] ?? $this->gatewayDistribute();
+
+        if(!$provider instanceof PaymentProviderInterface) {
+            $error = ['error'=>[__('Payment provider not available.')]];
+            return Helpers::error($error);
+        }
+
+        $baseMethod = $this->output['provider_method'] ?? $provider->defaultInitializeMethod();
+        $apiMethod = $baseMethod ? $baseMethod . 'Api' : null;
+
+        $response = $provider->initialize($this, [
+            'output' => $output,
+            'method' => $apiMethod,
+        ]);
         $output['response'] = $response;
-        if( $output['distribute'] == "pagaditoInitApi"){
+        if($apiMethod === 'pagaditoInitApi'){
             $parts = parse_url( $output['response']);
                 parse_str($parts['query'], $query);
                 // Extract the token value
