@@ -21,6 +21,7 @@ class Kernel extends HttpKernel
         \Illuminate\Foundation\Http\Middleware\ValidatePostSize::class,
         \App\Http\Middleware\TrimStrings::class,
         \Illuminate\Foundation\Http\Middleware\ConvertEmptyStringsToNull::class,
+        \App\Http\Middleware\TelemetryMiddleware::class,
     ];
 
     /**
@@ -34,6 +35,7 @@ class Kernel extends HttpKernel
             \Illuminate\Cookie\Middleware\AddQueuedCookiesToResponse::class,
             \Illuminate\Session\Middleware\StartSession::class,
             \App\Http\Middleware\SessionHardening::class,
+            \App\Http\Middleware\SensitiveSessionBinding::class,
             \Illuminate\View\Middleware\ShareErrorsFromSession::class,
             \App\Http\Middleware\VerifyCsrfToken::class,
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
@@ -45,6 +47,7 @@ class Kernel extends HttpKernel
         'api' => [
             // \Laravel\Sanctum\Http\Middleware\EnsureFrontendRequestsAreStateful::class,
             'throttle:api',
+            \App\Http\Middleware\Api\EnsurePartnerSecurity::class,
             \Illuminate\Routing\Middleware\SubstituteBindings::class,
             \App\Http\Middleware\Api\HandleLocalization::class
         ],
@@ -90,6 +93,8 @@ class Kernel extends HttpKernel
         'merchant.google.two.factor'    => \App\Http\Middleware\Merchant\GoogleTwoFactor::class,
         'merchant.google.two.factor.api'    => \App\Http\Middleware\Merchant\GoogleTwoFactorApi::class,
         'agent.google.two.factor'    => \App\Http\Middleware\Agent\GoogleTwoFactor::class,
+        'partner.security' => \App\Http\Middleware\Api\EnsurePartnerSecurity::class,
+        'sensitive.session' => \App\Http\Middleware\SensitiveSessionBinding::class,
         'agent.google.two.factor.api'    => \App\Http\Middleware\Agent\GoogleTwoFactorApi::class,
         'api.version' => \App\Http\Middleware\ApiVersionResolver::class,
         'auth.api' => \App\Http\Middleware\ApiAuthenticator::class,
@@ -104,6 +109,6 @@ class Kernel extends HttpKernel
         'user.registration.permission'  => \App\Http\Middleware\User\RegistrationPermission::class,
         'agent.registration.permission'  => \App\Http\Middleware\Agent\RegistrationPermission::class,
         'merchant.registration.permission'  => \App\Http\Middleware\Merchant\RegistrationPermission::class,
-        'admin.audit' => \App\Http\Middleware\Admin\AdminAuditLogger::class,
+        'feature' => \App\Http\Middleware\EnsureFeatureIsEnabled::class,
     ];
 }
