@@ -3,26 +3,34 @@
 namespace App\Providers;
 
 use App\Constants\ExtensionConst;
+use App\Models\Transaction;
+use App\Observers\TransactionObserver;
 use App\Providers\Admin\ExtensionProvider;
+use App\Services\Deployment\CanaryReleaseManager;
+use App\Services\Deployment\FeatureToggle;
 use Illuminate\Pagination\Paginator;
 use Illuminate\Support\Facades\Http;
-use Illuminate\Support\ServiceProvider;
 use Illuminate\Support\Facades\Schema;
-use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\URL;
+use Illuminate\Support\Facades\Validator;
+use Illuminate\Support\ServiceProvider;
 
 ini_set('memory_limit','-1');
 ini_set('serialize_precision','-1');
 class AppServiceProvider extends ServiceProvider
 {
+    use LogsAudit;
+
     /**
      * Register any application services.
      *
      * @return void
-     */
+    */
     public function register()
     {
-        //
+        $this->app->singleton(\App\Services\FeatureToggle::class, function ($app) {
+            return new \App\Services\FeatureToggle();
+        });
     }
 
     /**
@@ -34,12 +42,22 @@ class AppServiceProvider extends ServiceProvider
     {
         Paginator::useBootstrapFive();
         Schema::defaultStringLength(191);
-        if(config('security.enforce_https') && $this->app->environment('production') && !app()->runningInConsole()) {
+        if(config('app.force_https') && $this->app->environment('production') && !app()->runningInConsole()) {
             URL::forceScheme('https');
         }
 
         //laravel extend validation rules
         $this->extendValidationRule();
+
+        $this->logAuditAction('app_service_provider.boot', [
+            'payload' => [
+                'force_https' => config('app.force_https'),
+            ],
+            'result' => [
+                'validators_extended' => true,
+            ],
+            'status' => 'success',
+        ]);
     }
 
     /**
