@@ -6,10 +6,13 @@ use App\Models\Admin\Currency;
 use App\Models\AgentLoginLog;
 use App\Models\AgentWallet;
 use App\Models\DeviceFingerprint;
+use App\Traits\Security\LogsSecurityEvents;
 use Exception;
 use Jenssegers\Agent\Agent;
 
 trait LoggedInUsers {
+
+    use LogsSecurityEvents;
 
     protected function refreshUserWallets($user) {
         $user_wallets = $user->wallet->pluck("currency_id")->toArray();
@@ -59,8 +62,22 @@ trait LoggedInUsers {
 
         try{
             AgentLoginLog::create($data);
+            $this->logSecurityInfo('agent_login_success', [
+                'agent_id' => $user->id,
+                'fingerprint_id' => $fingerprint?->id,
+                'ip' => $client_ip,
+                'city' => $data['city'],
+                'country' => $data['country'],
+                'browser' => $data['browser'],
+                'os' => $data['os'],
+                'context' => 'agent_web',
+            ]);
         }catch(Exception $e) {
-            // return false;
+            $this->logSecurityError('agent_login_log_failed', [
+                'agent_id' => $user->id,
+                'ip' => $client_ip,
+                'message' => $e->getMessage(),
+            ]);
         }
     }
 }
