@@ -11,10 +11,16 @@
     <link href="https://fonts.googleapis.com/css2?family=Outfit:wght@100;200;300;400;500;600;700;800;900&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap" rel="stylesheet">
 
     @include('user.partials.header-assets')
+    @vite(['resources/css/app.css', 'resources/sass/app.scss', 'resources/js/app.js'])
 
     @stack('css')
 </head>
-<body class="{{ selectedLangDir() ?? "ltr"}}">
+@php
+    $user = auth()->user();
+    $themePreference = $user->preferred_theme ?? 'light';
+    $languagePreference = $user->preferred_language ?? app()->getLocale();
+@endphp
+<body class="{{ selectedLangDir() ?? "ltr"}}" data-theme="{{ $themePreference }}" data-language="{{ $languagePreference }}">
 
     <!--~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         Start body overlay
