@@ -1,9 +1,5 @@
 @extends('user.layouts.master')
 
-@push('css')
-
-@endpush
-
 @section('breadcrumb')
     @include('user.components.breadcrumb',['breadcrumbs' => [
         [
@@ -17,7 +13,10 @@
 <div class="body-wrapper">
     <div class="dashboard-area mt-10">
         <div class="dashboard-header-wrapper">
-            <h3 class="title">{{__(@$page_title)}}</h3>
+            <div class="d-flex align-items-center gap-2">
+                <h3 class="title mb-0">{{__(@$page_title)}}</h3>
+                @include('user.components.help-icon', ['section' => 'payments'])
+            </div>
         </div>
     </div>
     <div class="row mb-30-none">
@@ -80,73 +79,6 @@
                     <div class="dash-payment-body">
                         <div class="preview-list-wrapper">
 
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-coins"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{ __("Entered Amount") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="fw-bold request-amount">--</span>
-                                </div>
-                            </div>
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-battery-half"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{ __("Total Fees") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="fees">--</span>
-                                </div>
-                            </div>
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-receipt"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{ __("Recipient Received") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="recipient-get">--</span>
-                                </div>
-                            </div>
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-money-check-alt"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{__("Total Payable")}}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="last payable-total text-warning">--</span>
-                                </div>
-                            </div>
-                        </div>
-
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
     <div class="dashboard-list-area mt-20">
         <div class="dashboard-header-wrapper">
             <h4 class="title ">{{__("Make Payment Log")}}</h4>
@@ -161,242 +93,4 @@
         </div>
     </div>
 </div>
-<div class="modal fade" id="scanModal" tabindex="-1" role="dialog" aria-hidden="true">
-    <div class="modal-dialog modal-dialog-centered" role="document">
-      <div class="modal-content">
-            <div class="modal-body text-center">
-                <video id="preview" class="p-1 border" style="width:300px;"></video>
-            </div>
-            <div class="modal-footer justify-content-center">
-              <button type="button" class="btn btn-secondary btn-sm" data-bs-dismiss="modal">@lang('close')</button>
-            </div>
-      </div>
-    </div>
-</div>
 @endsection
-
-@push('script')
-<script src="https://rawgit.com/schmich/instascan-builds/master/instascan.min.js"></script>
-<script>
-//  'use strict'
-    (function ($) {
-        $('.scan').click(function(){
-            var scanner = new Instascan.Scanner({ video: document.getElementById('preview'), scanPeriod: 5, mirror: false });
-            scanner.addListener('scan',function(content){
-                var route = '{{url('user/merchant/qr/scan/')}}'+'/'+content
-                $.get(route, function( data ) {
-                    if(data.error){
-                        // alert(data.error)
-                        throwMessage('error',[data.error]);
-                    } else {
-                        $("#username").val(data);
-                        $("#username").focus()
-                    }
-                    $('#scanModal').modal('hide')
-                });
-            });
-
-            Instascan.Camera.getCameras().then(function (cameras){
-                if(cameras.length>0){
-                    $('#scanModal').modal('show')
-                        scanner.start(cameras[0]);
-                } else{
-                    throwMessage('error',["No camera found "]);
-                }
-            }).catch(function(e){
-                throwMessage('error',["No camera found "]);
-            });
-        });
-        $('.checkUser').on('keyup',function(e){
-            var url = '{{ route('user.make.payment.check.exist') }}';
-            var value = $(this).val();
-            var token = '{{ csrf_token() }}';
-            if ($(this).attr('name') == 'email') {
-                var data = {email:value,_token:token}
-
-            }
-            $.post(url,data,function(response) {
-                if(response.own){
-                    if($('.exist').hasClass('text--success')){
-                        $('.exist').removeClass('text--success');
-                    }
-                    $('.exist').addClass('text--danger').text(response.own);
-                    $('.transfer').attr('disabled',true)
-                    return false
-                }
-                if(response['data'] != null){
-                    if($('.exist').hasClass('text--danger')){
-                        $('.exist').removeClass('text--danger');
-                    }
-                    $('.exist').text(`Valid merchant for transaction.`).addClass('text--success');
-                    $('.transfer').attr('disabled',false)
-                } else {
-                    if($('.exist').hasClass('text--success')){
-                        $('.exist').removeClass('text--success');
-                    }
-                    $('.exist').text('Merchant doesn\'t  exists.').addClass('text--danger');
-                    $('.transfer').attr('disabled',true)
-                    return false
-                }
-
-            });
-        });
-    })(jQuery);
-</script>
-<script>
-     var defualCurrency = "{{ get_default_currency_code() }}";
-     var defualCurrencyRate = "{{ get_default_currency_rate() }}";
-
-    $(document).ready(function(){
-
-            getLimit();
-            getFees();
-            getPreview();
-        });
-        $("input[name=amount]").keyup(function(){
-             getFees();
-             getPreview();
-
-        });
-        $("input[name=amount]").focusout(function(){
-             enterLimit();
-
-        });
-        function getLimit() {
-            var currencyCode = acceptVar().currencyCode;
-            var currencyRate = acceptVar().currencyRate;
-
-            var min_limit = acceptVar().currencyMinAmount;
-            var max_limit =acceptVar().currencyMaxAmount;
-            if($.isNumeric(min_limit) || $.isNumeric(max_limit)) {
-                var min_limit_calc = parseFloat(min_limit/currencyRate).toFixed(2);
-                var max_limit_clac = parseFloat(max_limit/currencyRate).toFixed(2);
-                $('.limit-show').html("{{ __('limit') }} " + min_limit_calc + " " + currencyCode + " - " + max_limit_clac + " " + currencyCode);
-
-                return {
-                    minLimit:min_limit_calc,
-                    maxLimit:max_limit_clac,
-                };
-            }else {
-                $('.limit-show').html("--");
-                return {
-                    minLimit:0,
-                    maxLimit:0,
-                };
-            }
-        }
-        function acceptVar() {
-            var selectedVal = $("select[name=currency] :selected");
-            var currencyCode = $("select[name=currency] :selected").val();
-            var currencyRate = defualCurrencyRate;
-            var currencyMinAmount ="{{getAmount($makePaymentCharge->min_limit)}}"
-            var currencyMaxAmount = "{{getAmount($makePaymentCharge->max_limit)}}"
-            var currencyFixedCharge = "{{getAmount($makePaymentCharge->fixed_charge)}}"
-            var currencyPercentCharge = "{{getAmount($makePaymentCharge->percent_charge)}}"
-
-            return {
-                currencyCode:currencyCode,
-                currencyRate:currencyRate,
-                currencyMinAmount:currencyMinAmount,
-                currencyMaxAmount:currencyMaxAmount,
-                currencyFixedCharge:currencyFixedCharge,
-                currencyPercentCharge:currencyPercentCharge,
-                selectedVal:selectedVal,
-
-            };
-        }
-        function feesCalculation() {
-            var currencyCode = acceptVar().currencyCode;
-            var currencyRate = acceptVar().currencyRate;
-            var sender_amount = $("input[name=amount]").val();
-            sender_amount == "" ? (sender_amount = 0) : (sender_amount = sender_amount);
-
-            var fixed_charge = acceptVar().currencyFixedCharge;
-            var percent_charge = acceptVar().currencyPercentCharge;
-            if ($.isNumeric(percent_charge) && $.isNumeric(fixed_charge) && $.isNumeric(sender_amount)) {
-                // Process Calculation
-                var fixed_charge_calc = parseFloat(currencyRate * fixed_charge);
-                var percent_charge_calc = parseFloat(currencyRate)*(parseFloat(sender_amount) / 100) * parseFloat(percent_charge);
-                var total_charge = parseFloat(fixed_charge_calc) + parseFloat(percent_charge_calc);
-                total_charge = parseFloat(total_charge).toFixed(2);
-                // return total_charge;
-                return {
-                    total: total_charge,
-                    fixed: fixed_charge_calc,
-                    percent: percent_charge,
-                };
-            } else {
-                // return "--";
-                return false;
-            }
-        }
-
-        function getFees() {
-            var currencyCode = acceptVar().currencyCode;
-            var percent = acceptVar().currencyPercentCharge;
-            var charges = feesCalculation();
-            if (charges == false) {
-                return false;
-            }
-            $(".fees-show").html("{{ __('Payment Fee') }}: " + parseFloat(charges.fixed).toFixed(2) + " " + currencyCode + " + " + parseFloat(charges.percent).toFixed(2) + "%  ");
-        }
-        function getPreview() {
-                var senderAmount = $("input[name=amount]").val();
-                var sender_currency = acceptVar().currencyCode;
-                var sender_currency_rate = acceptVar().currencyRate;
-                senderAmount == "" ? senderAmount = 0 : senderAmount = senderAmount;
-                // Sending Amount
-                $('.request-amount').text(senderAmount + " " + defualCurrency);
-
-                // Fees
-                var charges = feesCalculation();
-                var total_charge = 0;
-                if(senderAmount == 0){
-                    total_charge = 0;
-                }else{
-                    total_charge = charges.total;
-                }
-
-                $('.fees').text(total_charge + " " + sender_currency);
-                // // recipient received
-                var recipient = parseFloat(senderAmount) * parseFloat(sender_currency_rate)
-                var recipient_get = 0;
-                if(senderAmount == 0){
-                     recipient_get = 0;
-                }else{
-                     recipient_get =  parseFloat(recipient);
-                }
-                $('.recipient-get').text(parseFloat(recipient_get).toFixed(2) + " " + sender_currency);
-
-                 // Pay In Total
-                var totalPay = parseFloat(senderAmount) * parseFloat(sender_currency_rate)
-                var pay_in_total = 0;
-                if(senderAmount == 0){
-                     pay_in_total = 0;
-                }else{
-                     pay_in_total =  parseFloat(totalPay) + parseFloat(charges.total);
-                }
-                $('.payable-total').text(parseFloat(pay_in_total).toFixed(2) + " " + sender_currency);
-
-        }
-        function enterLimit(){
-        var min_limit = parseFloat("{{getAmount($makePaymentCharge->min_limit)}}");
-        var max_limit =parseFloat("{{getAmount($makePaymentCharge->max_limit)}}");
-        var currencyRate = acceptVar().currencyRate;
-        var sender_amount = parseFloat($("input[name=amount]").val());
-
-        if( sender_amount < min_limit ){
-            throwMessage('error',['{{ __("Please follow the mimimum limit") }}']);
-            $('.transfer').attr('disabled',true)
-        }else if(sender_amount > max_limit){
-            throwMessage('error',['{{ __("Please follow the maximum limit") }}']);
-            $('.transfer').attr('disabled',true)
-        }else{
-            $('.transfer').attr('disabled',false)
-        }
-
-       }
-
-</script>
-
-@endpush
