@@ -1,9 +1,5 @@
 @extends('user.layouts.master')
 
-@push('css')
-
-@endpush
-
 @section('breadcrumb')
     @include('user.components.breadcrumb',['breadcrumbs' => [
         [
@@ -17,7 +13,10 @@
 <div class="body-wrapper">
     <div class="dashboard-area mt-10">
         <div class="dashboard-header-wrapper">
-            <h3 class="title">{{__($page_title)}}</h3>
+            <div class="d-flex align-items-center gap-2">
+                <h3 class="title mb-0">{{__($page_title)}}</h3>
+                @include('user.components.help-icon', ['section' => 'withdrawals'])
+            </div>
         </div>
     </div>
     <div class="row mb-30-none">
@@ -60,121 +59,6 @@
                                 </div>
                                 <div class="col-xl-6 col-lg-6 form-group">
 
-                                    <label>{{ __("Amount") }}<span>*</span></label>
-                                    <div class="input-group">
-                                        <input type="text" class="form--control" required placeholder="{{ __('enter Amount') }}" name="amount" value="{{ old("amount") }}">
-                                        <select class="form--control nice-select">
-                                            <option value="{{ get_default_currency_code() }}">{{ get_default_currency_code() }}</option>
-                                        </select>
-                                    </div>
-                                    <code class="d-block mt-10 text-end text--dark fw-bold balance-show">{{ __("Available Balance") }} {{ authWalletBalance() }} {{ get_default_currency_code() }}</code>
-                                </div>
-                                <div class="col-xl-12 col-lg-12 form-group">
-                                    <div class="note-area">
-                                        <code class="d-block limit-show">--</code>
-                                        <code class="d-block fees-show">--</code>
-                                    </div>
-                                </div>
-                                <div class="col-xl-12 col-lg-12">
-                                    <button type="submit" class="btn--base w-100 btn-loading">{{ __("Withdraw Money") }} <i class="fas fa-arrow-alt-circle-right ms-1"></i></button>
-                                </div>
-                            </div>
-                        </form>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="col-lg-6 mb-30">
-            <div class="dash-payment-item-wrapper">
-                <div class="dash-payment-item active">
-                    <div class="dash-payment-title-area">
-                        <span class="dash-payment-badge">!</span>
-                        <h5 class="title">{{ __($page_title) }} {{__("Preview")}}</h5>
-                    </div>
-                    <div class="dash-payment-body">
-                        <div class="preview-list-wrapper">
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-receipt"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{ __("Entered Amount") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="request-amount">--</span>
-                                </div>
-                            </div>
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="lab la-get-pocket"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{ __("Conversion Amount") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="conversionAmount">--</span>
-                                </div>
-                            </div>
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-battery-half"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{ __("Total Fees & Charges") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="fees">--</span>
-                                </div>
-                            </div>
-
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-money-check-alt"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span class="">{{ __("Will Get") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="text--success will-get">--</span>
-                                </div>
-                            </div>
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-money-check-alt"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span class="last">{{ __("Payable Amount") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="text--warning last total-pay">--</span>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
     <div class="dashboard-list-area mt-20">
         <div class="dashboard-header-wrapper">
             <h4 class="title ">{{__("withdraw Log")}}</h4>
@@ -190,188 +74,3 @@
     </div>
 </div>
 @endsection
-
-@push('script')
-    <script>
-         var defualCurrency = "{{ get_default_currency_code() }}";
-         var defualCurrencyRate = "{{ get_default_currency_rate() }}";
-
-        $('select[name=gateway]').on('change',function(){
-            getExchangeRate();
-            isCrypto();
-            getLimit();
-            getFees();
-            activeItems();
-        });
-        $(document).ready(function(){
-            getExchangeRate();
-            isCrypto();
-            getLimit();
-            getFees();
-            activeItems();
-        });
-        $("input[name=amount]").keyup(function(){
-             getFees();
-             activeItems();
-        });
-        function isCrypto(){
-            var pricison = 0;
-            var crypto = acceptVar().selectedVal.data('crypto');
-            if(crypto == 1){
-                 pricison = 8;
-            }else{
-                 pricison = 2;
-            }
-            return pricison;
-        }
-        function getExchangeRate() {
-            if(acceptVar().selectedVal.val() === "null"){
-                return false;
-            }
-            var pricison = isCrypto();
-            var currencyCode = acceptVar().currencyCode;
-            var currencyRate = acceptVar().currencyRate;
-            var currencyMinAmount = acceptVar().currencyMinAmount;
-            var currencyMaxAmount = acceptVar().currencyMaxAmount;
-            $('.rate-show').html("1 " + defualCurrency + " = " + parseFloat(currencyRate).toFixed(pricison) + " " + currencyCode);
-        }
-        function getLimit() {
-            var pricison = isCrypto();
-            var sender_currency = acceptVar().currencyCode;
-            var sender_currency_rate = acceptVar().currencyRate;
-            var min_limit = acceptVar().currencyMinAmount;
-            var max_limit =acceptVar().currencyMaxAmount;
-            if($.isNumeric(min_limit) || $.isNumeric(max_limit)) {
-                var min_limit_calc = parseFloat(min_limit/sender_currency_rate).toFixed(pricison);
-                var max_limit_clac = parseFloat(max_limit/sender_currency_rate).toFixed(pricison);
-                $('.limit-show').html("{{ __('limit') }} " + min_limit_calc + " " + defualCurrency + " - " + max_limit_clac + " " + defualCurrency);
-                return {
-                    minLimit:min_limit_calc,
-                    maxLimit:max_limit_clac,
-                };
-            }else {
-                $('.limit-show').html("--");
-                return {
-                    minLimit:0,
-                    maxLimit:0,
-                };
-            }
-        }
-        function acceptVar() {
-            var selectedVal = $("select[name=gateway] :selected");
-            var currencyCode = $("select[name=gateway] :selected").attr("data-currency");
-            var currencyRate = $("select[name=gateway] :selected").attr("data-rate");
-            var currencyMinAmount = $("select[name=gateway] :selected").attr("data-min_amount");
-            var currencyMaxAmount = $("select[name=gateway] :selected").attr("data-max_amount");
-            var currencyFixedCharge = $("select[name=gateway] :selected").attr("data-fixed_charge");
-            var currencyPercentCharge = $("select[name=gateway] :selected").attr("data-percent_charge");
-
-            // var sender_select = $("input[name=from_wallet_id] :selected");
-
-            return {
-                currencyCode:currencyCode,
-                currencyRate:currencyRate,
-                currencyMinAmount:currencyMinAmount,
-                currencyMaxAmount:currencyMaxAmount,
-                currencyFixedCharge:currencyFixedCharge,
-                currencyPercentCharge:currencyPercentCharge,
-                selectedVal:selectedVal,
-
-            };
-        }
-        function feesCalculation() {
-            var pricison = isCrypto();
-            var sender_currency = acceptVar().currencyCode;
-            var sender_currency_rate = acceptVar().currencyRate;
-            var sender_amount = $("input[name=amount]").val();
-            sender_amount == "" ? (sender_amount = 0) : (sender_amount = sender_amount);
-            var conversion_amount = parseFloat( sender_amount) *  parseFloat(sender_currency_rate)
-
-            var fixed_charge = acceptVar().currencyFixedCharge;
-            var percent_charge = acceptVar().currencyPercentCharge;
-            if ($.isNumeric(percent_charge) && $.isNumeric(fixed_charge) && $.isNumeric(conversion_amount)) {
-                // Process Calculation
-                var fixed_charge_calc = parseFloat(fixed_charge);
-                var percent_charge_calc = (parseFloat(conversion_amount) / 100) * parseFloat(percent_charge);
-                var total_charge = parseFloat(fixed_charge_calc) + parseFloat(percent_charge_calc);
-                total_charge = parseFloat(total_charge).toFixed(pricison);
-                // return total_charge;
-                return {
-                    total: total_charge,
-                    fixed: fixed_charge_calc,
-                    percent: percent_charge,
-                };
-            } else {
-                // return "--";
-                return false;
-            }
-        }
-        function getFees() {
-            var pricison = isCrypto();
-            var sender_currency = acceptVar().currencyCode;
-            var percent = acceptVar().currencyPercentCharge;
-            var charges = feesCalculation();
-            if (charges == false) {
-                return false;
-            }
-            $(".fees-show").html("{{ __('charge') }}: " + parseFloat(charges.fixed).toFixed(pricison) + " " + sender_currency + " + " + parseFloat(charges.percent).toFixed(pricison) + "%");
-        }
-        function activeItems(){
-            var selectedVal = acceptVar().selectedVal.val();
-            if(selectedVal === undefined || selectedVal === '' || selectedVal === null){
-                return false;
-            }else{
-                return getPreview();
-            }
-        }
-        function getPreview() {
-            var pricison = isCrypto();
-            var senderAmount = $("input[name=amount]").val();
-            var sender_currency = acceptVar().currencyCode;
-
-            if(acceptVar().selectedVal.val() === "null"){
-                return false;
-            }
-            var sender_currency_rate = acceptVar().currencyRate;
-            // var receiver_currency = acceptVar().rCurrency;
-            (senderAmount == "" || isNaN(senderAmount)) ? senderAmount = 0 : senderAmount = senderAmount;
-
-            // Sending Amount
-            $('.request-amount').text(senderAmount + " " + defualCurrency);
-
-            // Fees
-            var charges = feesCalculation();
-            var total_charge = 0;
-            if(senderAmount == 0){
-                total_charge = 0;
-            }else{
-                total_charge = charges.total;
-            }
-
-            $('.fees').text(total_charge + " " + sender_currency);
-
-            var conversionAmount = senderAmount * sender_currency_rate;
-            $('.conversionAmount').text(parseFloat(conversionAmount).toFixed(pricison) + " " + sender_currency);
-            // willget
-            var will_get = parseFloat(senderAmount) * parseFloat(sender_currency_rate)
-            var will_get_total = 0;
-            if(senderAmount == 0){
-                    will_get_total = 0;
-            }else{
-                    will_get_total =  parseFloat(will_get) - parseFloat(charges.total);
-            }
-            $('.will-get').text(parseFloat(will_get_total).toFixed(pricison) + " " + sender_currency);
-
-            // total payable
-            var totalPay = parseFloat(senderAmount)
-            var pay_in_total = 0;
-            if(senderAmount == 0){
-                    pay_in_total = 0;
-            }else{
-                //  pay_in_total =  parseFloat(totalPay) + parseFloat(charges.total);
-                    pay_in_total =  parseFloat(totalPay);
-            }
-            $('.total-pay').text(parseFloat(pay_in_total).toFixed(pricison) + " " + defualCurrency);
-        }
-    </script>
-@endpush
