@@ -1,9 +1,5 @@
 @extends('user.layouts.master')
 
-@push('css')
-
-@endpush
-
 @section('breadcrumb')
     @include('user.components.breadcrumb',['breadcrumbs' => [
         [
@@ -15,138 +11,80 @@
 
 @section('content')
 <div class="body-wrapper">
-    <div class="dashboard-area mt-10">
-        <div class="dashboard-header-wrapper">
-            <h3 class="title">{{__(@$page_title)}}</h3>
-        </div>
-    </div>
-    <div class="row mb-30-none">
-        <div class="col-xl-6 mb-30">
-            <div class="dash-payment-item-wrapper">
-                <div class="dash-payment-item active">
-                    <div class="dash-payment-title-area">
-                        <span class="dash-payment-badge">!</span>
-                        <h5 class="title">{{ __(@$page_title) }}</h5>
-                    </div>
-                    <div class="dash-payment-body">
-                        <form class="card-form" action="{{ setRoute('user.send.money.confirmed') }}" method="POST">
-                            @csrf
-                            <div class="row">
-                                <div class="col-xl-12 col-lg-12 form-group text-center">
-                                    <div class="exchange-area">
-                                        <code class="d-block text-center"><span class="fees-show">--</span> <span class="limit-show">--</span></code>
-                                    </div>
-                                </div>
-                                <div class="col-xxl-6 col-xl-12 col-lg-6 form-group paste-wrapper">
-                                    <label>{{ __("email Address") }} ({{ __("User") }})<span class="text--base">*</span></label>
-                                    <div class="input-group">
-                                        <div class="input-group-prepend">
-                                            <span class="input-group-text copytext"><span>{{ __("Email") }}</span></span>
-                                        </div>
-                                        <input type="email" name="email" class="form--control checkUser" id="username" placeholder="{{ __("enter Email Address") }}" value="{{ old('email') }}" />
-                                    </div>
-                                    <button type="button" class="paste-badge scan"  data-toggle="tooltip" title="Scan QR"><i class="fas fa-camera"></i></button>
-                                    <label class="exist text-start"></label>
+    @php
+        $minLimit = getAmount(($sendMoneyCharge->min_limit ?? 0) * get_default_currency_rate(), 2);
+        $maxLimit = getAmount(($sendMoneyCharge->max_limit ?? 0) * get_default_currency_rate(), 2);
+        $wizardConfig = [
+            'flow' => __('Exchange'),
+            'title' => __(@$page_title),
+            'subtitle' => __('Guide teammates through exchanging money between users with inline validation.'),
+            'formId' => 'send-money-wizard',
+            'csrfToken' => csrf_token(),
+            'currency' => get_default_currency_code(),
+            'locale' => app()->getLocale(),
+            'startingValues' => [
+                'email' => old('email'),
+                'amount' => old('amount'),
+            ],
+            'meta' => [
+                'previewTitle' => __('Send preview'),
+                'helper' => __('Fees and recipient amount update automatically.'),
+                'labels' => [
+                    'amount' => __('Amount'),
+                    'conversion' => __('Conversion'),
+                    'fees' => __('Fees & Charges'),
+                    'willGet' => __('Recipient will get'),
+                    'total' => __('Total Payable'),
+                    'limit' => __('Limit'),
+                    'rate' => __('Rate'),
+                ],
+            ],
+            'steps' => [
+                [
+                    'title' => __('Recipient details'),
+                    'description' => __('Confirm who should receive the funds'),
+                    'fields' => [
+                        [
+                            'name' => 'email',
+                            'type' => 'email',
+                            'label' => __('Recipient email'),
+                            'placeholder' => __('enter Email Address'),
+                            'required' => true,
+                            'helper' => __('We will detect if the user is eligible before submitting.'),
+                        ],
+                    ],
+                ],
+                [
+                    'title' => __('Send amount'),
+                    'description' => __('Define the amount you want to exchange'),
+                    'fields' => [
+                        [
+                            'name' => 'amount',
+                            'type' => 'number',
+                            'label' => __('Amount'),
+                            'placeholder' => __('enter Amount'),
+                            'required' => true,
+                            'rules' => [
+                                'min' => $minLimit,
+                                'max' => $maxLimit,
+                            ],
+                            'helper' => __('Respect the dynamic limits for your account.'),
+                        ],
+                    ],
+                ],
+                [
+                    'title' => __('Review & submit'),
+                    'description' => __('Confirm the payable amount and recipient share'),
+                    'fields' => [],
+                ],
+            ],
+        ];
+    @endphp
+    <div class="mt-4" data-flow-wizard='@json($wizardConfig)'></div>
+    <form id="send-money-wizard" action="{{ setRoute('user.send.money.confirmed') }}" method="POST" class="d-none">
+        @csrf
+    </form>
 
-                                </div>
-
-                                <div class="col-xxl-6 col-xl-12 col-lg-6 form-group">
-                                    <label>{{ __("Amount") }}<span>*</span></label>
-                                    <div class="input-group">
-                                        <input type="text" class="form--control number-input" required placeholder="{{__('enter Amount')}}" name="amount" value="{{ old("amount") }}">
-                                        <select class="form--control nice-select currency" name="currency">
-                                            <option value="{{ get_default_currency_code() }}">{{ get_default_currency_code() }}</option>
-                                        </select>
-                                    </div>
-                                    <code class="d-block mt-10 text-end text--warning balance-show">{{ __("Available Balance") }} {{ authWalletBalance() }} {{ get_default_currency_code() }}</code>
-                                </div>
-
-                                <div class="col-xl-12 col-lg-12">
-                                    <button type="submit" class="btn--base w-100 btn-loading transfer">{{ __("Confirm Send") }} <i class="fas fa-paper-plane ms-1"></i></i></button>
-                                </div>
-                            </div>
-                        </form>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div class="col-xl-6 mb-30">
-            <div class="dash-payment-item-wrapper">
-                <div class="dash-payment-item active">
-                    <div class="dash-payment-title-area">
-                        <span class="dash-payment-badge">!</span>
-                        <h5 class="title">{{__("Send Money Preview")}}</h5>
-                    </div>
-                    <div class="dash-payment-body">
-                        <div class="preview-list-wrapper">
-
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-coins"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{ __("Entered Amount") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="fw-bold request-amount">--</span>
-                                </div>
-                            </div>
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-battery-half"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{ __("Transfer Fee") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="fees">--</span>
-                                </div>
-                            </div>
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-receipt"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{ __("Recipient Received") }}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="recipient-get">--</span>
-                                </div>
-                            </div>
-                            <div class="preview-list-item">
-                                <div class="preview-list-left">
-                                    <div class="preview-list-user-wrapper">
-                                        <div class="preview-list-user-icon">
-                                            <i class="las la-money-check-alt"></i>
-                                        </div>
-                                        <div class="preview-list-user-content">
-                                            <span>{{__("Total Payable")}}</span>
-                                        </div>
-                                    </div>
-                                </div>
-                                <div class="preview-list-right">
-                                    <span class="last payable-total text-warning">--</span>
-                                </div>
-                            </div>
-                        </div>
-
-                    </div>
-                </div>
-            </div>
-        </div>
-    </div>
     <div class="dashboard-list-area mt-20">
         <div class="dashboard-header-wrapper">
             <h4 class="title ">{{__("Send Money Log")}}</h4>
@@ -161,222 +99,4 @@
         </div>
     </div>
 </div>
-<div class="modal fade" id="scanModal" tabindex="-1" role="dialog" aria-hidden="true">
-    <div class="modal-dialog modal-dialog-centered" role="document">
-      <div class="modal-content">
-            <div class="modal-body text-center">
-                <video id="preview" class="p-1 border" style="width:300px;"></video>
-            </div>
-            <div class="modal-footer justify-content-center">
-              <button type="button" class="btn btn-secondary btn-sm" data-bs-dismiss="modal">@lang('close')</button>
-            </div>
-      </div>
-    </div>
-</div>
 @endsection
-
-@push('script')
-<script src="https://rawgit.com/schmich/instascan-builds/master/instascan.min.js"></script>
-<script>
-//  'use strict'
-    (function ($) {
-        $('.scan').click(function(){
-            var scanner = new Instascan.Scanner({ video: document.getElementById('preview'), scanPeriod: 5, mirror: false });
-            scanner.addListener('scan',function(content){
-                var route = '{{url('user/qr/scan/')}}'+'/'+content
-                $.get(route, function( data ) {
-                    if(data.error){
-                        // alert(data.error)
-                        throwMessage('error',[data.error]);
-                    } else {
-                        $("#username").val(data);
-                        $("#username").focus()
-                    }
-                    $('#scanModal').modal('hide')
-                });
-            });
-
-            Instascan.Camera.getCameras().then(function (cameras){
-                if(cameras.length>0){
-                    $('#scanModal').modal('show')
-                        scanner.start(cameras[0]);
-                } else{
-                //    alert('No cameras found.');
-                    throwMessage('error',["No camera found "]);
-                }
-            }).catch(function(e){
-                // alert('No cameras found.');
-                throwMessage('error',["No camera found "]);
-            });
-        });
-        $('.checkUser').on('keyup',function(e){
-            var url = '{{ route('user.send.money.check.exist') }}';
-            var value = $(this).val();
-            var token = '{{ csrf_token() }}';
-            if ($(this).attr('name') == 'email') {
-                var data = {email:value,_token:token}
-
-            }
-            $.post(url,data,function(response) {
-                if(response.own){
-                    if($('.exist').hasClass('text--success')){
-                        $('.exist').removeClass('text--success');
-                    }
-                    $('.exist').addClass('text--danger').text(response.own);
-                    $('.transfer').attr('disabled',true)
-                    return false
-                }
-                if(response['data'] != null){
-                    if($('.exist').hasClass('text--danger')){
-                        $('.exist').removeClass('text--danger');
-                    }
-                    $('.exist').text(`Valid user for transaction.`).addClass('text--success');
-                    $('.transfer').attr('disabled',false)
-                } else {
-                    if($('.exist').hasClass('text--success')){
-                        $('.exist').removeClass('text--success');
-                    }
-                    $('.exist').text('User doesn\'t  exists.').addClass('text--danger');
-                    $('.transfer').attr('disabled',true)
-                    return false
-                }
-
-            });
-        });
-    })(jQuery);
-</script>
-<script>
-     var defualCurrency = "{{ get_default_currency_code() }}";
-     var defualCurrencyRate = "{{ get_default_currency_rate() }}";
-
-        $(document).ready(function(){
-
-            getLimit();
-            getFees();
-            getPreview();
-        });
-        $("input[name=amount]").keyup(function(){
-             getFees();
-             getPreview();
-        });
-        function getLimit() {
-            var currencyCode = acceptVar().currencyCode;
-            var currencyRate = acceptVar().currencyRate;
-
-            var min_limit = acceptVar().currencyMinAmount;
-            var max_limit =acceptVar().currencyMaxAmount;
-            if($.isNumeric(min_limit) || $.isNumeric(max_limit)) {
-                var min_limit_calc = parseFloat(min_limit/currencyRate).toFixed(2);
-                var max_limit_clac = parseFloat(max_limit/currencyRate).toFixed(2);
-                $('.limit-show').html("{{ __('limit') }} " + min_limit_calc + " " + currencyCode + " - " + max_limit_clac + " " + currencyCode);
-
-                return {
-                    minLimit:min_limit_calc,
-                    maxLimit:max_limit_clac,
-                };
-            }else {
-                $('.limit-show').html("--");
-                return {
-                    minLimit:0,
-                    maxLimit:0,
-                };
-            }
-        }
-        function acceptVar() {
-            var selectedVal = $("select[name=currency] :selected");
-            var currencyCode = $("select[name=currency] :selected").val();
-            var currencyRate = defualCurrencyRate;
-            var currencyMinAmount ="{{getAmount($sendMoneyCharge->min_limit)}}"
-            var currencyMaxAmount = "{{getAmount($sendMoneyCharge->max_limit)}}"
-            var currencyFixedCharge = "{{getAmount($sendMoneyCharge->fixed_charge)}}"
-            var currencyPercentCharge = "{{getAmount($sendMoneyCharge->percent_charge)}}"
-
-            return {
-                currencyCode:currencyCode,
-                currencyRate:currencyRate,
-                currencyMinAmount:currencyMinAmount,
-                currencyMaxAmount:currencyMaxAmount,
-                currencyFixedCharge:currencyFixedCharge,
-                currencyPercentCharge:currencyPercentCharge,
-                selectedVal:selectedVal,
-
-            };
-        }
-        function feesCalculation() {
-            var currencyCode = acceptVar().currencyCode;
-            var currencyRate = acceptVar().currencyRate;
-            var sender_amount = $("input[name=amount]").val();
-            sender_amount == "" ? (sender_amount = 0) : (sender_amount = sender_amount);
-
-            var fixed_charge = acceptVar().currencyFixedCharge;
-            var percent_charge = acceptVar().currencyPercentCharge;
-            if ($.isNumeric(percent_charge) && $.isNumeric(fixed_charge) && $.isNumeric(sender_amount)) {
-                // Process Calculation
-                var fixed_charge_calc = parseFloat(currencyRate * fixed_charge);
-                var percent_charge_calc = parseFloat(currencyRate)*(parseFloat(sender_amount) / 100) * parseFloat(percent_charge);
-                var total_charge = parseFloat(fixed_charge_calc) + parseFloat(percent_charge_calc);
-                total_charge = parseFloat(total_charge).toFixed(2);
-                // return total_charge;
-                return {
-                    total: total_charge,
-                    fixed: fixed_charge_calc,
-                    percent: percent_charge,
-                };
-            } else {
-                // return "--";
-                return false;
-            }
-        }
-
-        function getFees() {
-            var currencyCode = acceptVar().currencyCode;
-            var percent = acceptVar().currencyPercentCharge;
-            var charges = feesCalculation();
-            if (charges == false) {
-                return false;
-            }
-            $(".fees-show").html("{{ __('Transfer Fee') }}: " + parseFloat(charges.fixed).toFixed(2) + " " + currencyCode + " + " + parseFloat(charges.percent).toFixed(2) + "%  ");
-        }
-        function getPreview() {
-                var senderAmount = $("input[name=amount]").val();
-                var sender_currency = acceptVar().currencyCode;
-                var sender_currency_rate = acceptVar().currencyRate;
-                senderAmount == "" ? senderAmount = 0 : senderAmount = senderAmount;
-                // Sending Amount
-                $('.request-amount').text(senderAmount + " " + defualCurrency);
-
-                // Fees
-                var charges = feesCalculation();
-                var total_charge = 0;
-                if(senderAmount == 0){
-                    total_charge = 0;
-                }else{
-                    total_charge = charges.total;
-                }
-
-                $('.fees').text(total_charge + " " + sender_currency);
-                // // recipient received
-                var recipient = parseFloat(senderAmount) * parseFloat(sender_currency_rate)
-                var recipient_get = 0;
-                if(senderAmount == 0){
-                     recipient_get = 0;
-                }else{
-                     recipient_get =  parseFloat(recipient);
-                }
-                $('.recipient-get').text(parseFloat(recipient_get).toFixed(2) + " " + sender_currency);
-
-                 // Pay In Total
-                var totalPay = parseFloat(senderAmount) * parseFloat(sender_currency_rate)
-                var pay_in_total = 0;
-                if(senderAmount == 0){
-                     pay_in_total = 0;
-                }else{
-                     pay_in_total =  parseFloat(totalPay) + parseFloat(charges.total);
-                }
-                $('.payable-total').text(parseFloat(pay_in_total).toFixed(2) + " " + sender_currency);
-
-        }
-
-</script>
-
-@endpush
