@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Support\HelpContentController;
 use App\Http\Controllers\User\RequestMoneyController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -22,6 +23,7 @@ use App\Http\Controllers\User\SendMoneyController;
 use App\Http\Controllers\User\ReceipientController;
 use App\Http\Controllers\User\MobileTopupController;
 use App\Http\Controllers\User\PaymentLinkController;
+use App\Http\Controllers\User\PricingController;
 use App\Http\Controllers\User\TransactionController;
 use App\Http\Controllers\User\VirtualcardController;
 use App\Http\Controllers\User\ReceiveMoneyController;
@@ -29,11 +31,17 @@ use App\Http\Controllers\User\StripeVirtualController;
 use App\Http\Controllers\User\StrowalletVirtualController;
 use App\Http\Controllers\User\SudoVirtualCardController;
 use App\Http\Controllers\User\SupportTicketController;
+use App\Http\Controllers\User\PreferencesController;
 
 
+
+Route::get('/help/{section}', [HelpContentController::class, 'show'])
+    ->where('section', '[A-Za-z0-9\-]+')
+    ->name('help.show');
 
 Route::prefix("user")->name("user.")->group(function(){
     Route::post("info",[GlobalController::class,'userInfo'])->name('info');
+    Route::post('pricing/quote', [PricingController::class, 'quote'])->name('pricing.quote');
     Route::controller(DashboardController::class)->group(function(){
         Route::get('dashboard','index')->name('dashboard');
         Route::get('qr/scan/{qr_code}','qrScan')->name('qr.scan');
@@ -41,6 +49,9 @@ Route::prefix("user")->name("user.")->group(function(){
         Route::get('merchant/qr/scan/{qr_code}','merchantQrScan')->name('merchant.qr.scan');
         Route::post('logout','logout')->name('logout');
         Route::delete('delete/account','deleteAccount')->name('delete.account')->middleware('app.mode');
+    });
+    Route::controller(PreferencesController::class)->prefix('preferences')->name('preferences.')->group(function(){
+        Route::post('/', 'update')->name('update');
     });
     //profile
     Route::controller(ProfileController::class)->prefix("profile")->name("profile.")->middleware('app.mode')->group(function(){
@@ -139,11 +150,12 @@ Route::prefix("user")->name("user.")->group(function(){
     });
     //withdraw money
     Route::middleware('module:withdraw-money')->group(function(){
-        Route::controller(MoneyOutController::class)->prefix('withdraw')->name('money.out.')->group(function(){
+        Route::controller(MoneyOutController::class)->prefix('withdraw')->name('money.out.')->middleware('domain.rate_limit:withdrawal,user-withdraw')->group(function(){
             Route::get('/','index')->name('index');
             Route::post('insert','paymentInsert')->name('insert')->middleware('kyc.verification.guard');
             Route::get('preview','preview')->name('preview');
             Route::post('confirm','confirmMoneyOut')->name('confirm')->middleware('kyc.verification.guard');
+            Route::post('quote','quote')->name('quote');
 
             //check bank validation
             Route::post('check/flutterwave/bank','checkBanks')->name('check.flutterwave.bank');
@@ -156,7 +168,7 @@ Route::prefix("user")->name("user.")->group(function(){
     Route::middleware('module:virtual-card')->group(function(){
         //virtual card flutterwave
         Route::middleware('virtual_card_method:flutterwave')->group(function(){
-            Route::controller(VirtualcardController::class)->prefix('virtual-card')->name('virtual.card.')->group(function(){
+            Route::controller(VirtualcardController::class)->prefix('virtual-card')->name('virtual.card.')->middleware('domain.rate_limit:card,user-virtual-card')->group(function(){
                 Route::get('/','index')->name('index');
                 Route::post('create','cardBuy')->name('create')->middleware('kyc.verification.guard');
                 Route::post('fund','cardFundConfirm')->name('fund.confirm')->middleware('kyc.verification.guard');
@@ -169,7 +181,7 @@ Route::prefix("user")->name("user.")->group(function(){
         });
         //virtual card sudo
         Route::middleware('virtual_card_method:sudo')->group(function(){
-            Route::controller(SudoVirtualCardController::class)->prefix('sudo-virtual-card')->name('sudo.virtual.card.')->group(function(){
+            Route::controller(SudoVirtualCardController::class)->prefix('sudo-virtual-card')->name('sudo.virtual.card.')->middleware('domain.rate_limit:card,user-sudo-card')->group(function(){
                 Route::get('/','index')->name('index');
                 Route::post('create','cardBuy')->name('create')->middleware('kyc.verification.guard');
                 Route::post('make/default/remove/default','makeDefaultOrRemove')->name('make.default.or.remove');
@@ -181,7 +193,7 @@ Route::prefix("user")->name("user.")->group(function(){
         });
         //virtual card stripe
         Route::middleware('virtual_card_method:stripe')->group(function(){
-            Route::controller(StripeVirtualController::class)->prefix('stripe-virtual-card')->name('stripe.virtual.card.')->group(function(){
+            Route::controller(StripeVirtualController::class)->prefix('stripe-virtual-card')->name('stripe.virtual.card.')->middleware('domain.rate_limit:card,user-stripe-card')->group(function(){
                 Route::get('/','index')->name('index');
                 Route::post('create','cardBuy')->name('create')->middleware('kyc.verification.guard');
                 Route::get('details/{card_id}','cardDetails')->name('details');
@@ -193,7 +205,7 @@ Route::prefix("user")->name("user.")->group(function(){
         });
          //virtual card strowallet
          Route::middleware('virtual_card_method:strowallet')->group(function(){
-            Route::controller(StrowalletVirtualController::class)->prefix('strowallet-virtual-card')->name('strowallet.virtual.card.')->group(function(){
+            Route::controller(StrowalletVirtualController::class)->prefix('strowallet-virtual-card')->name('strowallet.virtual.card.')->middleware('domain.rate_limit:card,user-strowallet-card')->group(function(){
                 Route::get('/','index')->name('index');
 
                 Route::get('create','createPage')->name('create')->middleware('kyc.verification.guard');
@@ -213,14 +225,14 @@ Route::prefix("user")->name("user.")->group(function(){
     });
     //bill pay
     Route::middleware('module:bill-pay')->group(function(){
-        Route::controller(BillPayController::class)->prefix('bill-pay')->name('bill.pay.')->group(function(){
+        Route::controller(BillPayController::class)->prefix('bill-pay')->name('bill.pay.')->middleware('domain.rate_limit:payment,user-bill-pay')->group(function(){
             Route::get('/','index')->name('index');
             Route::post('insert','payConfirm')->name('confirm')->middleware('kyc.verification.guard');
         });
     });
     //Mobile TopUp
     Route::middleware('module:mobile-top-up')->group(function(){
-        Route::controller(MobileTopupController::class)->prefix('mobile-topup')->name('mobile.topup.')->group(function(){
+        Route::controller(MobileTopupController::class)->prefix('mobile-topup')->name('mobile.topup.')->middleware('domain.rate_limit:topup,user-mobile-topup')->group(function(){
             Route::get('/','index')->name('index');
             Route::post('type','selectType')->name('type');
             //manual methods
